create useGenerateThumbnail hook and improve thumnail fetching on client side
