+++ packages/app/app/studio/[organization]/nfts/create/components/SelectedMediaItem.tsx
@@ -6,7 +6,7 @@ import { IExtendedSession } from '@/lib/types';
 import { formatDate } from '@/lib/utils/time';
 import { XCircle } from 'lucide-react';
 import React from 'react';
-import { generateThumbnailAction } from '@/lib/actions/sessions';
+import useGenerateThumbnail from '@/lib/hooks/useGenerateThumbnail';
 
 const SelectedMediaItem = ({
   video,
@@ -15,19 +15,8 @@ const SelectedMediaItem = ({
   handleRemoveSelected: (video: IExtendedSession) => void;
   video: IExtendedSession;
 }) => {
-  const [generatedThumbnail, setGeneratedThumbnail] = useState<
-    string | undefined
-  >('');
-  const getThumbnail = async () => {
-    const thumbnail = await generateThumbnailAction(video);
+  const generatedThumbnail = useGenerateThumbnail({ session: video });
 
-    if (thumbnail) setGeneratedThumbnail(thumbnail);
-    return;
-  };
-  useEffect(() => {
-    if (video) getThumbnail();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [video]);
   return (
     <div className="relative mt-4">
       <div
